refactor: Restructure language support package
This patch splits the semantic analysis into multiple files, and moves
many of the interfaces and utility functions to new places for better
cohesion. Tests are updated accordingly, and some test cases are added
to cover gaps that became apparent after the restructuring.
