+++ packages/app/src/modules/editor/components/EditorPanel.tsx
@@ -1,10 +1,11 @@
+import type { SourceRange } from '@ast'
 import { syntaxTree } from '@codemirror/language'
 import type { Diagnostic } from '@codemirror/lint'
 import type { EditorView, ViewUpdate } from '@codemirror/view'
 import type { EditorLocation, PanelProps, Problem } from '@editor'
 import { Editor, getProjectFileContent, setProjectFileContent, useProjectSource, useProjectSourceDispatch, useProvideProblems } from '@editor'
 import type { RangeError } from '@language'
-import type { LanguageDiagnostic, SourceRange } from '@language-support'
+import type { LanguageDiagnostic } from '@language-support'
 import { applySemanticOperation, cadenceLanguageSupport, findUnusedVariables, goToDefinitionExtension, highlightOccurrencesExtension, hoverInfoExtension } from '@language-support'
 import type { FunctionComponent } from 'react'
 import { useCallback, useMemo, useState } from 'react'
