+++ packages/language-support/src/go-to-definition/extension.ts
@@ -3,9 +3,9 @@ import type { Extension } from '@codemirror/state'
 import { EditorSelection, StateEffect, StateField } from '@codemirror/state'
 import type { DecorationSet, ViewUpdate } from '@codemirror/view'
 import { Decoration, EditorView, ViewPlugin } from '@codemirror/view'
-import { sameRange } from '../analysis/model.js'
-import { applySemanticOperation } from '../operations.js'
-import type { SourceRange } from '../types.js'
+import { applySemanticOperation } from '../utilities/operations.js'
+import type { SourceRange } from '../utilities/range.js'
+import { sameRange } from '../utilities/range.js'
 import type { GoToDefinitionResult } from './operation.js'
 import { goToDefinition } from './operation.js'
