+++ packages/language-support/src/highlight-occurrences/operation.ts
@@ -1,6 +1,6 @@
-import { findReferenceRangesAt } from '../analysis/query.js'
-import type { SemanticOperation } from '../operations.js'
-import type { SourceRange } from '../types.js'
+import { findReferenceRangesAt } from '../model/query.js'
+import type { SemanticOperation } from '../utilities/operations.js'
+import type { SourceRange } from '../utilities/range.js'
 
 export const findHighlightedOccurrences: SemanticOperation<[pos: number], readonly SourceRange[]> = (model, pos) => {
   return findReferenceRangesAt(model, pos)
