+++ packages/language-support/src/utilities/diagnostic.ts
@@ -1,0 +1,7 @@
+import type { SourceRange } from './range.js'
+
+export interface LanguageDiagnostic {
+  readonly name: string
+  readonly message: string
+  readonly range: SourceRange
+}
