+++ packages/language-support/src/utilities/text.ts
@@ -1,4 +1,15 @@
-import type { SourceRange, TextLike } from '../types.js'
+import type { SourceRange } from './range.js'
+
+export interface TextLike {
+  readonly length: number
+  readonly sliceString: (from: number, to?: number) => string
+  readonly lineAt: (position: number) => TextLine
+}
+
+export interface TextLine {
+  readonly from: number
+  readonly number: number
+}
 
 export function textFromString (source: string): TextLike {
   const lineStarts = getLineStarts(source)
