--- packages/language-support/test/analysis/query.test.ts
@@ -1,147 +1,0 @@
-import type { Tree } from '@lezer/common'
-import { buildParser } from '@lezer/generator'
-import assert from 'node:assert'
-import { readFile } from 'node:fs/promises'
-import { describe, it } from 'node:test'
-import { analyzeTree } from '../../src/analysis/model.js'
-import { findDefinitionBindingAt } from '../../src/analysis/query.js'
-import { textFromString } from '../../src/analysis/text.js'
-import type { TextLike } from '../../src/types.js'
-import { getRangeAt } from '../helpers.js'
-
-const cadenceGrammar = await readFile(new URL('../../src/cadence.grammar', import.meta.url), 'utf8')
-const cadenceParser = buildParser(cadenceGrammar)
-
-function parseDocument (source: string): Readonly<{ tree: Tree, document: TextLike }> {
-  return {
-    tree: cadenceParser.parse(source),
-    document: textFromString(source)
-  }
-}
-
-describe('analysis/query.ts', () => {
-  it('resolves assignment bindings from the shared analysis query', () => {
-    const source = [
-      'kick = sample("/samples/kick.wav")',
-      'track (120.bpm) {',
-      '  part intro (4.bars) {',
-      '    kick << [x---]',
-      '  }',
-      '}',
-      ''
-    ].join('\n')
-
-    const { tree, document } = parseDocument(source)
-    const model = analyzeTree(tree, document)
-    const position = source.lastIndexOf('kick <<') + 1
-    const binding = findDefinitionBindingAt(model, position)
-
-    assert.ok(binding)
-    assert.strictEqual(binding.kind, 'assignment')
-    assert.strictEqual(binding.name, 'kick')
-
-    const kickOffset = source.indexOf('kick =')
-    assert.deepStrictEqual(binding.range, getRangeAt(source, kickOffset, 'kick'.length))
-  })
-
-  it('resolves explicit bus namespace access to the bus binding', () => {
-    const source = [
-      'track (120.bpm) {',
-      '  part foo (4.bars) {',
-      '    automate bus.foo.gain as curve [hold(0.db)]',
-      '  }',
-      '}',
-      'mixer {',
-      '  bus foo {}',
-      '}',
-      ''
-    ].join('\n')
-
-    const { tree, document } = parseDocument(source)
-    const model = analyzeTree(tree, document)
-    const position = source.indexOf('bus.foo.gain') + 'bus.'.length + 1
-    const binding = findDefinitionBindingAt(model, position)
-
-    assert.ok(binding)
-    assert.strictEqual(binding.kind, 'bus')
-    assert.strictEqual(binding.name, 'foo')
-
-    const busOffset = source.indexOf('bus foo') + 'bus '.length
-    assert.deepStrictEqual(binding.range, getRangeAt(source, busOffset, 'foo'.length))
-  })
-
-  it('does not resolve named argument keys', () => {
-    const source = [
-      'tempo = 128.bpm',
-      'track (tempo: 140.bpm) {}',
-      ''
-    ].join('\n')
-
-    const { tree, document } = parseDocument(source)
-    const model = analyzeTree(tree, document)
-    const position = source.indexOf('tempo:') + 1
-
-    assert.strictEqual(findDefinitionBindingAt(model, position), undefined)
-  })
-
-  it('does not resolve member access', () => {
-    const source = [
-      'use "effects" as fx',
-      'gain = 0.db',
-      'synth = sample("...", gain: gain)',
-      'track (120.bpm) {',
-      '  part intro (4.bars) {',
-      '    automate synth.gain as curve [hold(0.db)]',
-      '  }',
-      '}',
-      ''
-    ].join('\n')
-
-    const { tree, document } = parseDocument(source)
-    const model = analyzeTree(tree, document)
-    const position = source.indexOf('synth.gain') + 'synth.'.length + 1
-
-    assert.strictEqual(findDefinitionBindingAt(model, position), undefined)
-  })
-
-  it('does not resolve member access of an explicit bus access', () => {
-    const source = [
-      'track (120.bpm) {',
-      '  part main (4.bars) {',
-      '    automate bus.foo.gain as curve [hold(0.db)]',
-      '  }',
-      '}',
-      'mixer {',
-      '  bus foo {}',
-      '}',
-      ''
-    ].join('\n')
-
-    const { tree, document } = parseDocument(source)
-    const model = analyzeTree(tree, document)
-    const position = source.indexOf('bus.foo.gain') + 'bus.foo.'.length + 1
-
-    assert.strictEqual(findDefinitionBindingAt(model, position), undefined)
-  })
-
-  it('prefers import aliases over assignments with the same name', () => {
-    const source = [
-      'fx = sample("/samples/fx.wav")',
-      'use "effects" as fx',
-      'fx.delay()',
-      ''
-    ].join('\n')
-
-    const { tree, document } = parseDocument(source)
-    const model = analyzeTree(tree, document)
-    const position = source.lastIndexOf('fx.delay') + 1
-    const binding = findDefinitionBindingAt(model, position)
-
-    assert.ok(binding)
-    assert.strictEqual(binding.kind, 'use-alias')
-    assert.strictEqual(binding.name, 'fx')
-
-    const aliasOffset = source.indexOf('use "effects" as fx') + 'use "effects" as '.length
-    assert.deepStrictEqual(binding.range, getRangeAt(source, aliasOffset, 'fx'.length))
-  })
-})
